scripts/raw_data/requirements.txt
matplotlib
pyserial==3.5
